Create ConversationInfo folder, and import assets in RegisterSuccess instead of ussing path
