+++ src/pages/App/index.js
@@ -16,9 +16,6 @@ import {
   Text,
   IconButton,
   HStack,
-  VStack,
-  Button,
-  StackDivider,
 } from "@chakra-ui/react";
 import { Icon } from "@iconify-icon/react";
 import _ from "lodash";
@@ -35,7 +32,7 @@ import Message from "#/components/Message";
 import "./App.scss";
 import { SocketContext } from "#/contexts/SocketContext";
 import PreviewImageUpload from "#/components/PreviewImageUpload";
-
+import ConversationInfo from "./ConversationInfo";
 function App() {
   const { user, conversationId } = useContext(GlobalContext);
   const { socket, setSocket } = useContext(SocketContext);
@@ -275,159 +272,6 @@ function App() {
     handleRemoveImage,
   ]);
 
-  const [activeView, setActiveView] = useState("default");
-
-  const handleViewChange = (view) => {
-    setActiveView(view);
-  };
-  const renderConversationInfo = useMemo(() => {
-    return (
-      <VStack
-        width="30%"
-        height="100%"
-        padding="1rem 0"
-        borderLeft="solid lightgrey 1px"
-      >
-        {activeView === "default" && (
-          <VStack width={"100%"}>
-            <Box
-              width="100%"
-              display="flex"
-              flexDirection="column"
-              justifyContent="center"
-              alignItems="center"
-            >
-              <Avatar
-                size="xl"
-                src={getConversationAvatar(conversation, user.id)}
-                bg="gray.400"
-              />
-              <Text fontSize="lg" fontWeight="bold" marginTop="2">
-                {formatConversationName(conversation, user.id)}
-              </Text>
-            </Box>
-            <VStack
-              padding="0.5rem 0"
-              borderTop="solid lightgrey 1px"
-              divider={<StackDivider borderColor="gray.200" />}
-              width="100%"
-            >
-              <Button
-                onClick={() => handleViewChange("media")}
-                leftIcon={<Icon icon="pajamas:media" />}
-                justifyContent="flex-start"
-                width="100%"
-                variant="ghost"
-                color="black"
-                _hover={{ bg: "#ebedf0" }}
-              >
-                Media
-              </Button>
-              <Button
-                onClick={() => handleViewChange("file")}
-                leftIcon={<Icon icon="mdi:files" />}
-                justifyContent="flex-start"
-                width="100%"
-                variant="ghost"
-                color="black"
-                _hover={{ bg: "#ebedf0" }}
-              >
-                Files
-              </Button>
-              <Button
-                onClick={() => handleViewChange("link")}
-                leftIcon={<Icon icon="dashicons:admin-links" />}
-                justifyContent="flex-start"
-                width="100%"
-                variant="ghost"
-                color="black"
-                _hover={{ bg: "#ebedf0" }}
-              >
-                Links
-              </Button>
-            </VStack>
-          </VStack>
-        )}
-        ;
-        {activeView === "media" && (
-          <VStack width={"100%"}>
-            <HStack
-              position="relative"
-              display="flex"
-              justifyContent="center"
-              width="100%"
-              alignItems="center"
-            >
-              <Box
-                position="absolute"
-                left="0"
-                onClick={() => handleViewChange("default")}
-                padding="0.5rem"
-              >
-                <Icon icon="humbleicons:arrow-go-back" width="1.5rem" />
-              </Box>
-
-              <Text fontSize="lg" fontWeight="bold">
-                Media
-              </Text>
-            </HStack>
-          </VStack>
-        )}
-        ;
-        {activeView === "link" && (
-          <VStack width={"100%"}>
-            <HStack
-              position="relative"
-              display="flex"
-              justifyContent="center"
-              width="100%"
-              alignItems="center"
-            >
-              <Box
-                position="absolute"
-                left="0"
-                onClick={() => handleViewChange("default")}
-                padding="0.5rem"
-              >
-                <Icon icon="humbleicons:arrow-go-back" width="1.5rem" />
-              </Box>
-
-              <Text fontSize="lg" fontWeight="bold">
-                Links
-              </Text>
-            </HStack>
-          </VStack>
-        )}
-        ;
-        {activeView === "file" && (
-          <VStack width={"100%"}>
-            <HStack
-              position="relative"
-              display="flex"
-              justifyContent="center"
-              width="100%"
-              alignItems="center"
-            >
-              <Box
-                position="absolute"
-                left="0"
-                onClick={() => handleViewChange("default")}
-                padding="0.5rem"
-              >
-                <Icon icon="humbleicons:arrow-go-back" width="1.5rem" />
-              </Box>
-
-              <Text fontSize="lg" fontWeight="bold">
-                Files
-              </Text>
-            </HStack>
-          </VStack>
-        )}
-        ;
-      </VStack>
-    );
-  }, [user, conversation, activeView]);
-
   return (
     <>
       {!conversationId ? (
@@ -441,7 +285,12 @@ function App() {
           <FeedLayout height="100%" title={renderTitle} width="70%">
             {renderFeedChild}
           </FeedLayout>
-          {renderConversationInfo}
+          <ConversationInfo
+            user={user}
+            conversation={conversation}
+            getConversationAvatar={getConversationAvatar}
+            formatConversationName={formatConversationName}
+          />
         </HStack>
       )}
     </>
